backend/backend/requirements.txt
Flask
pydantic
pytest
google-cloud-firestore
